Python/requirements.txt
pandas
google-cloud-bigquery
google-cloud-storage
kaggle
apache_beam
kagglehub
db-dtypes
dash
plotly
